feat/cli: Added the --extended flag to command list_of_sports
This change allows the user to add the --extended option as a flag so that all the sports, including the minority sports are shown. The spyglass inner functionality has changed so that return two dicts: main and minority dict. Only in the CLI script will be decided which information will be shown.