feat: Corrected get_leagues_countries command in spyglass and CLI
This change engobles the TODO section that pointed out that it was being mixed countries and some competitions. Now it extracts the JS data and gets two dictionaries that are the ones according to the left part of the web. I have also made changes in the CLI, allowing the --expanded flag.