fix: Functionality get_reg_leagues corrected to obtain a dictionary
Previously this function provided the official name of the regional leagues, now it provides the keyname of the league as the key of the dictionary it returns. The value of that key is the official name. The necessary changes in the CLI have been made so that it uses the new dictionary. The documentation of the function has also been updated